# src/icast_cv/models/image.py
"""Image collection Pydantic models."""

from datetime import datetime

from pydantic import BaseModel, Field


class ImageCreate(BaseModel):
    """Internal model for creating an image document (not an API request body)."""

    sample_id: str
    filename: str
    file_path: str
    thumbnail_path: str
    width: int
    height: int
    file_size_bytes: int
    camera_index: int
    metadata: dict[str, object] = Field(default_factory=dict)
    captured_at: datetime


class ImageInDB(BaseModel):
    """Internal representation of an image document from MongoDB."""

    id: str
    sample_id: str
    filename: str
    file_path: str
    thumbnail_path: str
    width: int
    height: int
    file_size_bytes: int
    camera_index: int
    metadata: dict[str, object]
    captured_at: datetime


class ImageResponse(BaseModel):
    """API response for an image."""

    id: str
    sample_id: str
    filename: str
    file_path: str
    thumbnail_path: str
    width: int
    height: int
    file_size_bytes: int
    camera_index: int
    metadata: dict[str, object]
    captured_at: datetime